requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# CORS Headers
django-cors-headers>=4.3.0

# Environment Variables
python-decouple>=3.8
python-dotenv>=1.0.0

# AI/ML
google-generativeai>=0.8.0

# Utilities
Pillow>=10.0.0
pytz>=2023.3
python-dateutil>=2.8.0
PyPDF2>=3.0.0

# ASGI Server
daphne>=4.0.0
django-guardian==3.2.0
django-filter
